Viva corrections 7: Review references to Boltanski and Chiapello
Finally, we would like you to check your references properly, and give the
thesis a final proof read before submission. **There are some references
missing, or with inconsistent dates** (for example, Boltanski and Chiapello),
which probably just results from the edition you consulted at the time, but it
is important to include all cited sources in the reference list.