Viva corrections 6: Clarify the use of the terms “civic” and “public”
In one or two places (eg p 58) you refer to the civic. How do you distinguish
this from the 'public'? This could be an important distinction in political
theory, as the civic refers specifically to the 'city' as the space of
political action. If this distinction is not important, you might simply want
to stick to public, though 'civic' does open a nice connection to the 'cities'
in Boltanski and Chiapello's ideas about political ideologies (and the
projective city of the new spirit).