Viva corrections 1: Conceptualise “incumbent institution”
We would like you to clearly define and conceptualise the concept of an
'incumbent institution' in your thesis. The best place for this is probably in
the Literature Review, as this really functions as a theoretical framework for
your thesis. We are not expecting you to go into institutional theory in doing
this but you might like to consider exactly what you mean by an institution,
when this covers both nation states and markets. This discussion of
institutions might also consider how institutions and ideology are connected,
given your use of Mouffe and Boltanski and Chiapello later in the thesis