Viva corrections 3: Include short discussion on platform capitalism
Incorporate a short discussion to position your thesis in relation to the
literature on the political economy of platform capitalism more broadly. The
best place for this might be in the introduction. We are not asking for
another detailed literature review, just a brief overview of key debates and
issues to help locate your contribution in this academic literature. Suggested
sources for this would include Benkler's The Wealth of Networks and Nick
Srnicek's Platform Capitalism.

You might also want to consider some of Christian Fuchs' extensive body of
work on the political economy of social media, search engines, and other
elements of the platform economy. The choice of literature to include is up to
you really, but we both felt that some consideration of this wider literature,
at least to locate the distinctiveness of crowd-funding platforms, would
strengthen the thesis and open it up to another readership.