Viva corrections 5a: Review the use of the concept “money”
On page 73 you discuss the 'simplistic' understandings of money in Weber and
Marx. As discussed in the viva, this does not fairly represent either. Weber,
contra to your comment on page 109 about the apparently universal will to
accumulate, clearly analyses the concrete cultural and historical
particularities in which this will to accumulate first came into being (The
Protestant Ethic and the Spirit of Capitalism). Marx very clearly lays out a
duality of money (in Capital vol 1), distinguishing the money in the pocket of
a labourer from 'capital', which is only capital so long as it circulates and
accumulates. In this sense, your points about accumulation apply specifically
to capital, rather than money per se. This point should really be made clearer
in these parts of the thesis.