Viva corrections 5b: Review the use of the concept of “alienation”
On page 120 (and on) you have a short discussion of alienation that is unclear.
You quote Felipe, then mention Carla, in terms of job dissatisfaction, then say
that this is 'not a Marxist alienation' (p 120) but then go on to quote Marx
(p 121). This really needs a clearer account of Marx on alienation as there are
four aspects to this, as discussed in his 1844 Economic and Philosophical
Manuscripts: alienation of the worker from her products, her self, her
colleagues, and her species being.

In discussing the lack of control and creativity in her work, Felipe could be
understood as articulating a very Marxist conception of alienation from self-
directed activity and species being.

Whatever you choose to argue here, this point needs to be clarified and
conceptualised more seriously. This is not just an insignificant point of
minor theoretically precision. It is important because a move to alienation as
the core of the critique of capitalism is central to the 1968 movements that
Boltanski and Chiapello see as the basis for the new spirit of capitalism. If
crowd-funding is understood to be articulating a critique of capitalist work
on this same basis, then your subsequent use of Boltanski and Chiapello is
more clearly justified and stronger.