Viva corrections 4b: Discuss methodological strategies to handle divergent respondents
Connected to this we would also like a stronger discussion of the status of
divergent positions in your interviews. We discussed this in the viva, but it
is also noted in the thesis that different people articulated different
positions, and you structure the last two chapters around a kind of optimism/
pessimism perspective on the political potential of crowd-funding. Are these
positions connected to different respondents, and did you consider how
individual differences might explain these perspectives, or were they both
articulated by the same respondents?

This could be clearer in the main body of the thesis, but it also has
implications for how you coded and analysed the interviews. We would like to
see quite a bit more about coding in the methodology, both in terms of how you
actually did the coding and analysis, but also how you go beyond exactly what
was said in analysing the interviews.